refactor: useRuleFormModel 유틸 함수 분리 (#30)
