+++ src/pages/rules/hooks/useRuleFormModel.ts
@@ -1,8 +1,13 @@
-import { WEEK_DAYS_MON_FIRST, type WeekDay } from '@/constants/weekdays';
+import type { WeekDay } from '@/constants/weekdays';
 import {
   ruleCreateSchema,
   type RuleCreateValues,
 } from '@/schemas/ruleCreateSchema';
+import {
+  DEFAULT_RULE_TIME_RANGE,
+  parseRepeatDays,
+  toRepeatDays,
+} from '@/utils/ruleForm';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
@@ -36,36 +41,6 @@ const INITIAL_SETTINGS: RuleSettings = {
   repeatEnabled: false,
 };
 
-const DEFAULT_TIME_RANGE = {
-  startTime: '09:00:00',
-  endTime: '12:00:00',
-} as const;
-
-// UI 선택 요일 배열 -> "1,3,5" 문자열
-const toRepeatDays = (selectedDays: WeekDay[]) => {
-  return WEEK_DAYS_MON_FIRST.reduce<string[]>((acc, day, index) => {
-    if (selectedDays.includes(day)) {
-      acc.push(String(index + 1));
-    }
-
-    return acc;
-  }, []).join(',');
-};
-
-// "1,3,5" -> UI 선택 요일 배열
-const parseRepeatDays = (repeatDays: string) => {
-  if (!repeatDays.trim()) return [];
-
-  return repeatDays
-    .split(',')
-    .map((day) => Number.parseInt(day, 10))
-    .filter(
-      (dayNumber) =>
-        Number.isFinite(dayNumber) && dayNumber >= 1 && dayNumber <= 7
-    )
-    .map((dayNumber) => WEEK_DAYS_MON_FIRST[dayNumber - 1]);
-};
-
 const getInitialSettings = (
   initialValues?: RuleFormInitialValues
 ): RuleSettings => ({
@@ -76,18 +51,9 @@ const getInitialSettings = (
 });
 
 const getInitialTimeRange = (initialValues?: RuleFormInitialValues) => ({
-  startTime: initialValues?.startTime ?? DEFAULT_TIME_RANGE.startTime,
-  endTime: initialValues?.endTime ?? DEFAULT_TIME_RANGE.endTime,
+  startTime: initialValues?.startTime ?? DEFAULT_RULE_TIME_RANGE.startTime,
+  endTime: initialValues?.endTime ?? DEFAULT_RULE_TIME_RANGE.endTime,
 });
-
-export const formatKoreanTime = (time: string) => {
-  const [hourText, minuteText = '00'] = time.split(':');
-  const hour = Number.parseInt(hourText, 10);
-  const period = hour < 12 ? '오전' : '오후';
-  const displayHour = hour % 12 === 0 ? 12 : hour % 12;
-
-  return `${period} ${displayHour}:${minuteText}`;
-};
 
 type UseRuleFormModelParams = {
   initialValues?: RuleFormInitialValues;
@@ -148,8 +114,8 @@ export const useRuleFormModel = ({
     setActiveTimeField(null);
   };
 
+  // 화면 상태와 RHF 입력값을 API 요청 스펙에 맞춰 결합
   const handleFormSubmit = (data: RuleCreateValues) => {
-    // 화면 상태와 RHF 입력값을 API 요청 스펙에 맞춰 결합
     onSubmit({
       title: data.title,
       memo: data.memo ?? '',
